Fix fast refresh issue
